Split the setup-arch script into 2 scripts.
As explained in a comment in the setup-arch.sh file, making this split
is necessary because after we change roots to /mnt, the script will
waiting until the chroot is finished. Thus, I had to split the script in
such a way that whatever needs to be run inside the chroot'ed /mnt is run
in the "step2" of the setup.

It was used a very simple "trick" to make the chroot'ed system run the
step2 script: before chrooting, the step2 script is moved to /mnt/root
and the file /mnt/root/.bashrc is created just to call it. Then, when
chrooting, the script will be run. After everything is done, both the
step2 script and the .bashrc file are deleted.